Fixing compilation errors

+++ nms/src/nms_cuda.cpp
@@ -25,19 +25,19 @@ int gpu_nms(at::Tensor keep, at::Tensor num_out, at::Tensor boxes, float nms_ove
 
   const int col_blocks = DIVUP(boxes_num, threadsPerBlock);
   at::Tensor mask = at::empty(at::CUDA(at::kLong), {boxes_num, col_blocks});
-  long long* mask_flat = mask.data<long long>();
+  int64_t* mask_flat = mask.data<int64_t>();
 
   _nms(boxes_num, boxes_flat, mask_flat, nms_overlap_thresh);
 
   at::Tensor mask_cpu = mask.toBackend(at::Backend::CPU);
 
-  long long * mask_cpu_flat = mask_cpu.data<long long>();
+  int64_t * mask_cpu_flat = mask_cpu.data<int64_t>();
 
   at::Tensor remv_cpu = at::zeros(at::CPU(at::kLong), {col_blocks});
-  long long* remv_cpu_flat = remv_cpu.data<long long>();
+  int64_t* remv_cpu_flat = remv_cpu.data<int64_t>();
 
-  long long * keep_flat = keep.data<long long>();
-  long long num_to_keep = 0;
+  int64_t * keep_flat = keep.data<int64_t>();
+  int64_t num_to_keep = 0;
 
   int i, j;
   for (i = 0; i < boxes_num; i++) {
@@ -46,14 +46,14 @@ int gpu_nms(at::Tensor keep, at::Tensor num_out, at::Tensor boxes, float nms_ove
 
     if (!(remv_cpu_flat[nblock] & (1ULL << inblock))) {
       keep_flat[num_to_keep++] = i;
-      long long *p = &mask_cpu_flat[0] + i * col_blocks;
+      int64_t *p = &mask_cpu_flat[0] + i * col_blocks;
       for (j = nblock; j < col_blocks; j++) {
         remv_cpu_flat[j] |= p[j];
       }
     }
   }
 
-  long long * num_out_flat = num_out.data<long long>();
+  int64_t * num_out_flat = num_out.data<int64_t>();
   * num_out_flat = num_to_keep;
 
   return 1;
